Фикс попытки удалить несуществующую директорию

+++ text_to_video/logic.py
@@ -93,7 +93,8 @@ def main(text, outp = "outp.mp4", size = SIZE, fps = FPS, length = LENGH_IN_SECO
     if length == None:
         length = LENGH_IN_SECONDS
     base_img = BaseTextImage(text, "tnr.ttf").img
-    shutil.rmtree(frames_folder) 
+    if os.path.isdir(frames_folder):
+        shutil.rmtree(frames_folder) 
     RenderTools.part_render_loop(base_img, frames_folder, size, fps, length)
     return RenderTools.render_video(frames_folder,outp, size, fps)
